+++ tests/TestCase.php
@@ -138,7 +138,7 @@ GRAPHQL;
             | DebugFlag::RETHROW_UNSAFE_EXCEPTIONS,
         );
 
-        $config->set('lighthouse.guard', null);
+        $config->set('lighthouse.guards', null);
 
         $config->set('lighthouse.subscriptions', [
             'version' => 1,