+++ tests/Unit/Support/Http/Middleware/AttemptAuthenticationTest.php
@@ -25,7 +25,7 @@ final class AttemptAuthenticationTest extends TestCase
         $config->set('lighthouse.route.middleware', [
             AttemptAuthentication::class,
         ]);
-        $config->set('lighthouse.guard', 'foo');
+        $config->set('lighthouse.guards', ['foo']);
         $config->set('auth.guards.foo', [
             'driver' => 'foo',
             'provider' => 'users',